+++ src/fuselage/CCPACSFuselage.cpp
@@ -747,6 +747,34 @@ void CCPACSFuselage::SetFuselageHelper(CTiglFuselageHelper& cache) const
     cache.SetFuselage(const_cast<CCPACSFuselage*>(this));
 }
 
+void CCPACSFuselage::ReorderSections()
+{
+    auto sorted_element_uids = GetSegments().GetElementUIDsInOrder();
+
+    std::unordered_map<std::string, size_t> order;
+    order.reserve(sorted_element_uids.size());
+    for (size_t i = 0; i < sorted_element_uids.size(); ++i) {
+        order[sorted_element_uids[i]] = i;
+    }
+
+    auto get_rank = [&](auto const& section_ptr) -> size_t {
+        assert(section_ptr != nullptr); // Avoid exception, catch nullptr dereferencing in debug mode.
+        auto element_uid = section_ptr->GetElements().GetElement(1).GetUID();
+        auto it = order.find(element_uid);
+        if (it != std::end(order)) {
+            return it->second;
+        }
+        return std::numeric_limits<size_t>::max();
+    };
+
+    auto& sections = GetSections().GetSections();
+    std::stable_sort(
+        std::begin(sections),
+        std::end(sections),
+        [&](auto const& l, auto const& r) { return get_rank(l) < get_rank(r); }
+    );
+}
+
 void CCPACSFuselage::CreateNewConnectedElementBetween(std::string startElementUID, std::string endElementUID, double eta, std::string sectionName)
 {
     if(0.0001 > eta || eta > 0.9999)
@@ -791,6 +819,7 @@ void CCPACSFuselage::CreateNewConnectedElementBetween(std::string startElementUI
 
     // connect the element with segment and update old segment
     GetSegments().SplitSegment(segmentToSplit, newElement->GetSectionElementUID() );
+    ReorderSections();
 }
 
 std::optional<std::string> CCPACSFuselage::GetElementUIDAfterNewElement(std::string startElementUID)
@@ -823,7 +852,6 @@ void CCPACSFuselage::CreateNewConnectedElementAfter(std::string startElementUID,
     auto elementUIDAfter = GetElementUIDAfterNewElement(startElementUID);
     if (elementUIDAfter) {
         CreateNewConnectedElementBetween(startElementUID, *elementUIDAfter, 0.5, sectionName);
-        throw tigl::CTiglError("No eta is specified, but the new section lies between two sections.");
     }
     else {
         // in this case we simply need to find the previous element and call the appropriate function
@@ -909,7 +937,6 @@ void CCPACSFuselage::CreateNewConnectedElementBefore(std::string startElementUID
     auto elementUIDBefore = GetElementUIDBeforeNewElement(startElementUID);
     if (elementUIDBefore) {
         CreateNewConnectedElementBetween(*elementUIDBefore, startElementUID, 0.5, sectionName);
-        throw tigl::CTiglError("No eta is specified, but the new section lies between two sections.");
     }
     else {
         std::vector<std::string> elementsAfter  =  ListFunctions::GetElementsAfter(fuselageHelper->GetElementUIDsInOrder(), startElementUID);
@@ -969,6 +996,7 @@ void CCPACSFuselage::CreateNewConnectedElementBefore(std::string startElementUID
         if ( m_segments.NeedReordering() ){
             try { // we use a try-catch to not rise two time a exception if the reordering occurs during the first cpacs parsing
                 m_segments.ReorderSegments();
+                ReorderSections();
             }
             catch (  const CTiglError& err) {
                 LOG(ERROR) << err.what();